Complete 03-flex-header-2
It's not really clear if they are expecting me to not use the
margin: auto trick that's significant enough to warrant it's own
section in one of the first articles they list to read. Since it
achieves the same visual look at the desired outcome, I'm sticking
with it since it's one less div and class to deal with.